mira-bots/telegram_test_runner/requirements.txt
telethon>=1.36.0
Pillow>=10.0.0
PyYAML>=6.0
